client: Add a basic command line client for mogwai-scheduled
This adds mogwai-schedule-client, which is basically a simple wget which
supports scheduling. Before downloading the requested file, it creates a
schedule entry in the scheduler, and waits until the scheduler tells it
to start downloading.

The intention is that mogwai-schedule-client is used for simple testing
of the scheduler, rather than as a valuable scripting tool. At the
moment, it is useful enough to exercise most of the mogwai-scheduled
D-Bus API.

This adds a dependency on libsoup-2.4.

It includes documentation for the client, but no test cases.

Signed-off-by: Philip Withnall <withnall@endlessm.com>
https://phabricator.endlessm.com/T20421