daemon: Fix SystemCallFilter in systemd .service file
The previous filter was not permissive enough to let the daemon start.
After some experimentation, it turns out it’s not reasonable to
implement this as a whitelist: systemd is still adding more system call
groups to make this easier, and we would have to have listed all the
elements of @file-system (which comes from a newer version of systemd
than we have on EOS), plus others.

Instead, use a blacklist, and blacklist all the system call groups from
our version of systemd which we don’t need.

Secondly, set SystemCallArchitectures=native to ensure that syscalls
can’t be made using a different architecture to avoid the filtering.
This is useful on x86_64, where apparently syscalls can be made using
the x86 syscall API at any time.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T20422