libmogwai-schedule-client: Add a basic client library
This is the client side of the D-Bus API. gdbus-codegen hasn’t been used
here, since a nicer public C API can be written (only) by hand. The
mechanically generated gdbus-codegen APIs are not nice to use, and are
subject to change in future, so can’t make any API guarantees.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T20423