build: Use correct mogwai-scheduled API version for client
The client is named and versioned after the API version of
mogwai-scheduled, but it was using a copy of the API versioning variable
with the wrong value.

Eliminate the copied variable, fix the versioning; swap the subdir()
calls in the top-level meson.build around so the client can use the
variable from mogwai-scheduled.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T20423