libmogwai-tariff: Add library for reading/writing tariff files
This adds the libmogwai-tariff library, which allows descriptions of
network connection tariffs to be loaded and saved. A network tariff is a
collection of one or more periods where a fixed set of properties apply
to the connection, such as a capacity limit or a bandwidth limit.

At the moment, this will need to be connected to a UI for users to
describe their network tariffs individually. In the long-distant future,
it might be possible to ship a collection of pre-built tariffs for
common mobile networks.

The library is fully documented. Packaging and tests will follow
subsequently.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T20748