debian: Package mogwai-tariff and add tests
Package mogwai-tariff in a new mogwai-tariff-tools package, and add
autopkgtests for it, just to smoketest that it (and libmogwai-tariff)
work on an installed system.

At some point in the future it might make sense to move the autopkgtests
out to proper installed-tests, so they can be run by
gnome-desktop-testing-runner *or* by autopkgtest.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T20751