misc: Clean up use of g_auto(GVariantDict)
G_VARIANT_DICT_INIT does already exist, but not in the API reference I
was using, and the compiler doesn’t suggest using it unless you already
pass a parameter to it.

This tidies up the code a little, but introduces no functional changes.

It bumps the glib-2.0 dependency of mogwai-schedule-client, but that
makes no functional difference since other parts of Mogwai already
depend on a higher version of GLib (2.54).

Signed-off-by: Philip Withnall <withnall@endlessm.com>