libmogwai-tariff: Fix alignment fix for tariffs
The alignment fix introduced in commit e45e444a3b was, surprisingly,
correct — except that the aligned heap memory was promptly discarded and
the old unaligned memory chunk was passed to g_variant_new_from_bytes()
instead. That was not the intention of that commit. Oops.

Fix that and, this time, verify that the fix works using an ARM chroot.
This gives us the opportunity to check various other bits of the
compilation and unit tests on ARM, and extract some tariff files from
ARM which we can use to test the endianness correction between
architectures. So this isn’t a total loss.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T20396