libmogwai-schedule: Add a basic scheduling function
Move the existing hard-coded scheduling logic into a separate
mws_scheduler_reschedule() function, which is where all future
scheduling decisions will be done. Add an EntryData ancillary structure
for each entry in the scheduler, so we can track its scheduling state
and history. Add a signal which will be emitted when an entry’s
scheduling state changes.

The general idea here is that mws_scheduler_reschedule() will be called
whenever an input to the scheduler changes (entry added or removed,
network state changes, connection capacity usage changes, etc.), and it
will recalculate the scheduling state and update entries where
necessary.

Currently, the scheduling algorithm produces the same output for all
schedule entries: download iff the network is unmetered. That will
change in future.

mws_scheduler_reschedule() is exposed publicly so it can be called from
unit tests, but clients are not expected to need to call it themselves.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T21135