libmogwai-schedule: Add NetworkManager connection monitor
This isn’t plumbed in yet. Once it is, it will monitor the active
network connections and provide relevant information to the scheduler to
inform its scheduling decisions.

The connection monitor is provided as an interface and a NetworkManager
implementation so that dummy implementations of the interface can be
written for unit testing in future. This is the general architecture I
want to use to ensure that the scheduler itself remains unit testable
without too much pain.

This adds a dependency on libnm; at least version 1.8 (since that’s what
we have in EOS 3.3), and at most version 1.10 (since that’s what we have
in master).

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T21135