mogwai-scheduled: Plumb connection monitor in to the scheduler
Expand the scheduling algoritm to deal with the new network connection
information, modify the service setup to asynchronously construct a
NetworkManager monitor, and install a polkit rules file to grant
mogwai-scheduled permission to query connection settings from
NetworkManager.

With this commit, the combination of eos-updater and mogwai-scheduled is
almost at feature parity with the previous implementation of metered
data support in eos-updater. We just need to query the user’s
preferences about whether to download on a metered connection now.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T21135