libmogwai-schedule: Drop our internal hash table of NMDevices
Instead of keeping our own record of the currently available NMDevices
which are relevant to the NMActiveConnections (and ignoring the
NMDevices which are not relevant to any active connection), just track
the set of them using signals from the NMClient. Whenever an NMDevice
emits a signal we might care about, check whether it’s currently
associated with an active connection; if not, ignore the signal.

This simplifies management of devices without changing behaviour.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T21135