libmogwai-schedule: Drop our internal hash table of NMActiveConnections
Similarly to the previous commit, there’s actually no need to keep an
internal copy of the set of active connections, given that the NMClient
object has one already.

We still need to cache the array of connection IDs, since NMClient
doesn’t provide that. We also need to search through NMClient’s array of
active connections in order to find one by ID – rather than doing a hash
table lookup – but given that there are typically only a couple of
active connections (and almost never more than 10), this should not be a
problem.

This should introduce no observable behavioural changes.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T21135