libmogwai-schedule: Add tariff, download-when-metered connection details
Add two new fields to the MwtConnectionDetails struct: a tariff, and
whether to download from this connection when it’s metered. Implement
them in the NetworkManager implementation of MwsConnectionMonitor.

The values loaded from NetworkManager are user settings which are set by
gnome-control-center (and currently by nothing else).

In order to add the libmogwai-tariff dependency to libmogwai-schedule,
the subdir inclusion order in the top-level meson.build needed to be
shifted around.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T21135