libmogwai-tariff: Support recurrence in mwt_period_contains_time()
Add a lot of unit tests for it. This supports checking whether a given
date/time lies within an #MwtPeriod or any of its recurrences. It will
form a key part of scheduling: knowing which period applies to the
current time.

This also adds a new method, mwt_period_get_next_recurrence(), which
will get the start and end time of the next recurrence of a period, if
there is to be another recurrence. The implementations are fairly
tightly linked, as are the tests.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T21322