libmogwai-schedule: Automatically reschedule when a tariff transitions
When a tariff transitions between two periods, its scheduling properties
change, so the scheduler needs to do a reschedule then. Add basic
support to the scheduler to do this.

The support added here is very minimal: it does not properly handle the
differences between monotonic and wall clock time, and hence will run
into trouble if the user changes timezones or suspends their computer.
Future tasks will improve this behaviour, and add unit tests (which
can’t currently be done due to handling times far in the future). The
whole thing needs to be abstracted out.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T21322