libhelper: Store a GMainContext to attach GSources to
Rather than attaching the UNIX signal handler GSources to the default
main context, attach them to an explicit GMainContext, which we save
earlier as the thread default main context at the time of constructing
the HlpService object.

In practice, this makes no difference at the moment; it moves the time
at which the main context is bound from when hlp_service_run() is called
to when the HlpService object is constructed.

The stored GMainContext will become more useful in a subsequent commit,
where more GSources are attached to it at different times. In such a
situation, we definitely wouldn’t want to rely on the thread default
main context staying the same over time.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T20743