libhelper: Add inactivity timeout support to HlpService
This adds a ‘hold count’ to HlpService, and a new
HlpService:inactivity-timeout property. If the inactivity timeout is
reached and the hold count is zero, the service automatically exits.

Implementations of the service should increase the hold count while
handling a user request, or doing something else which the service
should remain running for. They should decrement the hold count
afterwards.

By default, the inactivity timeout is disabled, but it can be set by a
command line argument, or its default value changed by implementations
of HlpService.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T20743