tests: Add basic test for mogwai-scheduled --inactivity-timeout
This is a simple Python test harness which wraps mogwai-scheduled1 and
runs it as a subprocess to see if it will exit after its inactivity
timeout. We don’t test that it *doesn’t* exit if schedule entries are
added, as that should be done by the tests which test adding schedule
entries. (Those tests don’t exist yet, but are out of scope for adding
as part of this task.)

Add a Debian package for the tests, installed-test integration, and
ensure it’s run by sadt.

The Python is written so that python-dbusmock support could be added to
it in future to mock up (for example) the NetworkManager API on the
system bus for further testing.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T20743