mogwai-tariff: Convert basic Debian tests to installed-tests in Python
Following the example from the previous commit, convert the existing
tests for mogwai-tariff, which are in debian/tests/, to a full-blown
Python test harness in mogwai-tariff/tests/. The set of functionality
tested is equivalent, but the new version of it will be installed and
hence can be run by gnome-desktop-testing-runner. It will still be run
by sadt due to debian/tests/gnome-desktop-testing.

This eliminates some technical debt.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T20743