libmogwai-tariff: Support storing period timezones, v2 file format
Bump the file format version number to 2 and add support for storing the
timezone of the start and end of a period — previously these were always
in UTC. Being able to store the timezone is pretty important for making
sure that a period always starts/ends at the right time for parts of the
world which observe DST switches.

This also adds support for a ‘floating timezone’ (where the timezone is
set to the empty string), which is always instantiated as the user’s
current local timezone. This is not supported in the public API, yet,
however, as GDateTime has no way of representing it. Support may be
extended from the file format to the public API in future.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T21497