libmogwai-schedule: Allow limiting the number of active entries
Modify the scheduler to allow a limit on the number of active entries at
any time — i.e. a limit on the number of parallel downloads. This
requires an ordering to be applied over the entries which //could// be
active, in order to choose the most important ones to actually be
active. Currently, this ordering is ill-defined. Future commits will
define it better, based on the owner of the entry.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T21327