libmogwai-schedule: Use /proc/$pid/cmdline, not exe, for credentials
It turns out that you need to be root to readlink(/proc/$pid/exe), or to
have some combination of capabilities involving CAP_SYS_PTRACE which I
cannot get to work.

So, to get the credentials for a peer, read /proc/$pid/cmdline instead.
This is less secure (there are more races inherent in reading it and
resolving the resulting program name against $PATH; and processes can
modify their own cmdline to whatever they want), but since this is only
used for scheduling downloads, it’s not the end of the world if a
process forges its cmdline. Use of the download scheduler is not
mandatory, so if a process wants a high priority download, they can just
not ask the download scheduler and download it directly anyway.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T21327