connection-monitor-nm: Connect to NMConnection:changed
It was discovered that the correct way to respond to NetworkManager
changes in connections is by connecting to NMConnection:changed signal,
instead of each NMSetting implementation's notify.

This commit changes that and, because of the scope change, renames
the functions and data structures to properly reflect what is being
conected to and being fired.

https://phabricator.endlessm.com/T21823