libmogwai-tariff: Fix timezone serialisation for tariffs
The start and end times of a period in a tariff are stored as UNIX
timestamps and timezone abbreviations. It turns out that it’s not
possible to construct a GTimeZone instance from such an abbreviation,
meaning that timezone information was essentially lost in the round
trip.

Fix that by storing the timezone identifier instead of its abbreviation.
The identifier is what’s passed to g_time_zone_new() in the first place,
so should survive a round trip.

This relies on new API in GLib which allows us to access the timezone
and its identifier, which means we need to bump our dependency to
2.54.2-1endless3, which contains the backports from GLib master.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T21500