tariff-loader: Don't convert timezones
It is currently not possible to set a timezone that is
different from the GDateTime that was used to create
a MwtPeriod. If you`re in -0300 and creates a GDateTime
representing 01-05-2018 20:00, the UNIX timestamp is
always going to be binded to that timezone.

The tariff loader code, however, assumes that the
timezone and the UNIX timestamps are separated, and
converts the timestamp to the given timezone. That
causes the loaded period always different than the
saved period, on timezones different than UTC.

Fix that by not converting the UNIX timestamp from
UTC to the timezone, but rather creating the GDateTime
from the UNIX timestamp and assuming that timezone.

https://phabricator.endlessm.com/T22384