libmogwai-schedule: Emit active-entries-changed when removing entries
Previously, the MwsScheduler::active-entries-changed signal would only
be emitted when an entry is added and immediately becomes active. We
should also emit it when an active entry is removed from the scheduler.

This results in (potentially) two emissions of active-entries-changed
when calling update_entries(): one to remove the existing active
entries, and then one to add the new ones, with an emission of
entries-changed between.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T20423