tests: Add support for mock dummy credentials to MwsPeerManagerDummy
Expand the support in MwsPeerManagerDummy for returning mock data, so
that full peer credentials can be returned to callers if needed. This
will be used in upcoming commits to test the scheduler.

This introduces an internal API break, in that peers now need to be
explicitly added to the MwsPeerManagerDummy before any non-failure
results will be returned for them.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T20423