tests: Drop signal disconnection from MwsSignalLogger
If the MwsSignalLogger is freed before the GObject whose signals it’s
connected to, it would try to g_signal_handler_disconnect() its closure
in a closure invalidation notifier callback. However, GObject always
connects its own closure invalidation notifier callback which does the
same thing. This would result in one of the
g_signal_handler_disconnect() calls emitting a critical warning due to
being passed an invalid signal handler ID.

Leave the signal handler disconnection in that case to the GObject
callback. This means we can drop the weak reference to the GObject. We
could also drop MwsLoggedClosure.signal_id, but it’s useful for
debugging, so keep it around for that only.

In the case where the GObject whose signals are connected to is
finalised before the MwsSignalLogger, the behaviour remains unchanged.
Similarly in the case where the caller calls
g_signal_handler_disconnect() themselves using the signal handler ID
returned by mws_signal_logger_connect().

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T20423