tests: Handle two signals for active entry removal separately
Previously assert_entries_changed_signals() assumed that the signal for
removal of active entries was always emitted by itself before the
entries-changed signal, and then both followed by another
active-entries-changed signal which announced the new active entries.

However, that is not true in the case where an active entry becomes
inactive due to another (higher priority, say) entry becoming active and
inactivating it due to hitting the max-active-entries limit.

Add another argument to assert_entries_changed_signals() so that the two
types of removed entry can be listed separately by tests, and hence
correctly checked for in the right signal emission.

None of the existing tests expose this subtlety: it will be used in an
upcoming test.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T20423