build: Add glib-2.0 subproject
This is needed because we depend on GLib 2.57.1, which isn’t packaged
widely (or at all) yet. The subproject dependency can be dropped once
2.57.1 is available in the CI images we use.

This means we need to temporarily add more Debian build dependencies,
and disable --werror (since GLib is full of warnings with the
project-wide warning options we enable for Mogwai).

Signed-off-by: Philip Withnall <withnall@endlessm.com>