tests: Fix scheduler tests on 32-bit architectures
The scheduler tests pass a 64-bit vararg (capacity-limit) to
mwt_period_new(), but without explicit size indications, the compiler
will assume the constant is a normal integer. On 32-bit architectures
this means a 32-bit vararg is put in, and a 64-bit vararg is read out.
This gives erroneous results and was causing the scheduler tests to
fail.

Signed-off-by: Philip Withnall <withnall@endlessm.com>