mogwai-schedule-client: Add support for libsoup-3.0
The main change is the removal of the SoupRequest object and using the
soup_session_send_async API directly on the SoupSession instead.

This commit adds support for using libsoup-3.0 instead of libsoup-2.4,
and sets it as the default. libsoup-2.4 can still be used if Mogwai is
configured with `-Dsoup2=true`.

See
https://discourse.gnome.org/t/please-build-against-libsoup-3-by-default/10190
for context.