Revert "mogwai-scheduled: Fix dependency name for systemd"
This reverts commit ed84c69746e6aa3da963d0064032c9271dd7a221, and
additionally changes the CI config to install the package providing
`systemd.pc`, which was missing before. `libsystemd-dev` needs to be
kept as a dependency, as `libsystemd.so` is actually used by
the libgsystemservice subproject. The CI build uses libgsystemservice as
a subproject as it’s not available on Debian Unstable at the moment.

The pkg-config name was correct before. Both `libsystemd.pc` and
`systemd.pc` exist, but `systemd.pc` is the one which defines all
systemd’s directory paths and variables. `libsystemd.pc` is just for the
cflags/libs for linking to libsystemd.

Unsure how I managed to get this commit working before.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>