scheduler: Consider metered connections for allow-downloads.
This changes MwsScheduler to consider whether the connection is metered
when it is calculating its allow-downloads property. This allows for a
client application to know ahead of time if a download will be delayed
due to the connection being metered. This does not affect the behaviour
of download scheduling itself: `mws_scheduler_reschedule()` already
accounts for metered connections in the same way when it updates the
download schedule.

https://phabricator.endlessm.com/T35301